init epoch on changefeed creation
